minDiskFreePct should be a float (fixes #829)

+++ src/main/java/com/nutomic/syncthingandroid/model/Folder.java
@@ -17,7 +17,7 @@ public class Folder {
     public int rescanIntervalS;
     public final boolean ignorePerms = true;
     public boolean autoNormalize;
-    public int minDiskFreePct;
+    public float minDiskFreePct;
     public Versioning versioning;
     public int copiers;
     public int pullers;
